refactor: 이벤트 예외 클래스 분리 및 중복 제거
- EventException 중첩 클래스를 독립적인 파일로 분리
- EventNotFoundException, InvalidEventStatusException, InvalidEventTypeException을 독립 클래스로 생성
- EventExceptionHandler 삭제 (GlobalExceptionHandler와 중복)
- EventException.java 파일 삭제 (불필요한 중첩 구조 제거)
- 단일 책임 원칙 준수 및 코드 가독성 향상
